[chore] #114 코드 리뷰 수정

+++ app/src/main/java/org/sopt/dateroad/presentation/ui/past/PastScreen.kt
@@ -96,7 +96,7 @@ fun PastScreen(
                     PastCard(
                         date = pastUiState.dates[index],
                         dateType = DateType.getDateTypeByIndex(index),
-                        onClick = { navigateToTimelineDetail(DateType.getDateTypeByIndex(index), pastUiState.dates[index].dateId.toInt()) }
+                        onClick = { navigateToTimelineDetail(DateType.getDateTypeByIndex(index), pastUiState.dates[index].dateId) }
                     )
                 }
             }
